(refactor) Frames 객체 생성하여 ScoreBoard Frame 관련 분리
