+++ src/main/java/bowling/controller/BowlingController.java
@@ -1,5 +1,6 @@
 package bowling.controller;
 
+import bowling.domain.frame.Frames;
 import bowling.domain.frame.ScoreBoard;
 import bowling.domain.frame.dto.ScoreBoardAssembler;
 import bowling.domain.pin.Pin;
@@ -11,14 +12,15 @@ public class BowlingController {
 
     public void start() {
         User user = User.of(InputView.getUserName());
-        ScoreBoard scoreBoard = ScoreBoard.init(user);
+        Frames frames = Frames.init();
+        ScoreBoard scoreBoard = ScoreBoard.init(user, frames);
 
-        OutputView.print(ScoreBoardAssembler.assemble(scoreBoard));
+        OutputView.print(ScoreBoardAssembler.assemble(scoreBoard, frames));
 
         while (!scoreBoard.isGameOver()) {
-            OutputView.printCurrentFrame(scoreBoard.getLastIndex());
+            OutputView.printCurrentFrame(frames.getLastIndex());
             scoreBoard.bowl(Pin.of(InputView.getFelledPin()));
-            OutputView.print(ScoreBoardAssembler.assemble(scoreBoard));
+            OutputView.print(ScoreBoardAssembler.assemble(scoreBoard, frames));
         }
     }
 }
